Moved filetype plugins from /plugins to /ftplugins
This now causes only the plugin file for the current filetype
to be run, instead of running all the filetype plugins to be
run in alphabetical order. This fixes a bug caused by me not
reading the help page closely enough.

Also adds makeprgs for python and vim files. Python files are
now executed in the python interpreter on :make, and vim files
now do nothing on :make (set makeprg=true)